// src/app/shared/modules/api-client/model/newMasterForm.ts
/**
 * bsi-qf-backend
 * Formularios de Inspección de Calidad BSI
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { MasterFormBlockExcludingId } from './masterFormBlockExcludingId';


/**
 * (tsType: Omit<MasterForm, \'id\'>, schemaOptions: { title: \'NewMasterForm\', exclude: [ \'id\' ] })
 */
export interface NewMasterForm { 
    name: string;
    section: string;
    type: string;
    formato?: string;
    edicion?: string;
    blocks?: Array<MasterFormBlockExcludingId>;
}